Major re-desing/re-implementation of SwiftHLM
Also, in case HLM backend is not provided/configured a dummy connector/backend
provided within the SwiftHLM package will be used by default (it is very dummy
and needs to be improved)

*****  External interface changes *****

SwiftHLM version 0.2.1 provides the following basic HLM functions on the external
Swift interface:

- MIGRATE (container or an object from disk to HLM)
- RECALL (i.e. prefetch a container or an object from HLM to disk)
- STATUS (get status for a container or an object)
- REQUESTS (get status of migration and recall requests previously submitted
  for a contaner or an object).

-------
MIGRATE
-------

Trigger a migration from disk to HLM of a single object or all objects within a
container.
MIGRATE request is an HTTP POST request, with the following syntax:

    POST http://<host>:<port>/hlm/v1/MIGRATE/<account>/<container>/<object>
    POST http://<host>:<port>/hlm/v1/MIGRATE/<account>/<container>

    Note: SwiftHLM request keywords are case-insensitive, MIGRATE or migrate can
    be used, RECALL or recall, STATUS or status, REQUESTS or requests.

------
RECALL
------

Trigger a recall from HLM to disk for a single object or all objects within a
container.
RECALL request is an HTTP POST request, with the following syntax:

    POST http://<host>:<port>/hlm/v1/RECALL/<account>/<container>/<object>
    POST http://<host>:<port>/hlm/v1/RECALL/<account>/<container>

------
STATUS
------

Return, as the response body, a JSON encoded dictionary of objects and their
status (on HLM or on disk) for a given object or all objects within a
container.
STATUS query request is an HTTP GET request, with the following syntax:

    GET http://<host>:<port>/hlm/v1/STATUS/<account>/<container>/<object>
    GET http://<host>:<port>/hlm/v1/STATUS/<account>/<container>

------
REQUESTS
------

Return, as the response body, a JSON encoded list of pending or failed
migration and recall requests submitted for a contaner or an object.
REQUESTS query request is an HTTP GET request, with the following syntax:

    GET http://<host>:<port>/hlm/v1/REQUESTS/<account>/<container>/<object>
    GET http://<host>:<port>/hlm/v1/REQUESTS/<account>/<container>

E.g. to use SwiftHLM interfrace via a standard (unmodified) curl Swift client:

curl -H "X-Auth-Token: $TOKEN" -X POST "http://zagreb.zurich.ibm.com:8080/hlm/v1/migrate/AUTH_test/cont1
curl -H "X-Auth-Token: $TOKEN" -X POST "http://zagreb.zurich.ibm.com:8080/hlm/v1/recall/AUTH_test/cont1
curl -H "X-Auth-Token: $TOKEN" -X GET "http://zagreb.zurich.ibm.com:8080/hlm/v1/status/AUTH_test/cont1
curl -H "X-Auth-Token: $TOKEN" -X GET "http://zagreb.zurich.ibm.com:8080/hlm/v1/requests/AUTH_test/cont1

***** SwiftHLM - Backend interface changes *****

SwiftHLM Generic Backend API version 0.2.1

(3 digits API versions, such version as 0.2.1, should be considered
developmental and not stable version)

    response = SwiftHlmBackendConnector.submit_request_get_response(request)

request =
  {
  command : status,
  objects :
    [
      { object : /a/c/obj1, file : /srv/node/filepath1 },
      { object : /a/c/obj2, file : /srv/node/filepath2 }
    ]
  }

response =
  {
  objects :
    [
      {object : /a/c/obj1, file : /srv/node/filepath1, status : migrated,},
      {object : /a/c/obj2, file : /srv/node/filepath2, status : resident},
      {object : /a/c/obj3, file : /srv/node/filepath3, status : premigrated},
      {object : /a/c/obj4, file : /srv/node/filepath4, status : unknown}
    ]
  }

The deta structures used are dicitionary and list, the values are strings,
shown above unquoted and additinally indented for easier reading.

In addition to 'status', other requests are 'migrate' or 'recall' for which the
response is integer:
    0 - success
    1 - 1 or more objects could not be migrated/recalled
    2 - unable to process request for all objects (e.g. cannot invoke backend)

***** New Design/Internals Overview *****

This is a short overview of SwiftHLM components and their join operation for
providing SwiftHLM functionality.

SwiftHLM workflow for processing MIGRATION requests (and it is same for RECALL
requests) is as follows.  SwiftHLM middleware on proxy server intrcepts
SwiftHLM migration requests and queues them inside Swift, by storing them into
a special HLM-dedicated container as zero size objects.  After SwiftHLM request
is queued, 202 code is returned to the application.

Another SwiftHLM process, called SwiftHLM Dispatcher, is processing the queued
requests asynchronously with respect to user/application that submitted them.
It picks a request from the queue, in FIFO or a more advance manner, and groups
the requests into one list per involved storage node.

For each storage node/list Dispatcher invokes remotely a SwiftHLM program on
that storage node (the name of that program is SwiftHLM Handler), and provides
it with the list. Handler could also be a long running process listening for
and processing submissions from Dispatcher. Either way, the function performed
by Handler is to map the objects to files (or to HLM backend objects) and
submits the file list and the migration requests to HLM backend, if the backend
already provide the function to move data between LLM (low latency media) and
HLM (hight latency media). Examples of backends with such function are IBM
Spectrum Archive and BDT Tape Library Connector.

In order to support different backends, a Generic Backend Interface is defined
and used by Handler to submit the request to HLM backend, via the backend
specific Connector that maps the request to the backend specific API. If HLM
backend does not support moving data and managing object state, the backend
Connector needs to implement that function as well.

Once the backend completes the operation the result (succes or failure) is
propagated back to the dispatcher. In case of success, the request is removed
from the queue, otherwise it is marked as failed and kept in the queue for some
period (to be able to answer the request status queries). One could also
consider implementing request retries.

Querying object status (STATUS) is processed by SwiftHLM middleware
synchrounously, by groupping the queries per storage nodes and invoking the
Handler (same as Dispatcher does for migration and recall), but for status the
SwiftHLM middleware also merges the statuses reportd by the backend and
provides the merged result to the Swift application.

Querying requests status (REQUESTS) Query for requests status for an object or
a container are processed by SwiftHLM middleware, by reading listing of the
special HLM-dedicated container(s). If there are not pending (incompleted or
failed) requests for a container, the previously submitted operations for that
container may be considered completed. This is more efficient than to query
state for each object of a container.