Improve configuration file parsing.
While the previous code worked for most cases, for installations that use
non-standard parameters in the [DEFAULT] section of the Object server
configuration file, the file<->object mapping did not work due to inability
to get the DiskFile using a non-standard devices parameter in the config file.

Root cause is that swift.common.utils.readconf() seems to read the entire
config file, but uses RawConfigParser.sections() to retrieve the sections,
which does _not_ include the [DEFAULT] section.