Fix task queue object delimiter naming issue.
When using '--' as delimiter for the task queue object name within the hidden
.swifthlm account, containers or objects that included '--' broke the object
name parsing. The fixed implementation uses the '/' as delimiter and takes
care that objects containing '/' will also be parsed correctly.
Note that the change does not ensure backward compatibility, so make sure
to remove all pending (failed) requests from the account before upgrading.