Basic status caching using memcache.
Object status is cached in the proxy-local memcache, in the first place to
speed up GET operations, but also to improve STATUS query performance.
Note that upon migration and recall submission the cache is invalidated,
but if a STATUS operation is executed before migration and recall is
completed, the object status might be invalid after migration and recall
successfully completed. This needs to be fixed before merging this code
to the main branch.