Improved Unicode object path support.
On an Object get operation, presence of an object was checked in a way that did
not find objects whose object path contained unicode characters.
Also for status caching we need to take care that we construct the keys in a
unicode path compatible way to prevent charset errors for debug log output.